picasso/requirements.txt
Django
djangorestframework

celery[redis]

django-filter
envparse
pytest-django